% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_data_normalization.R
\name{data_normalization}
\alias{data_normalization}
\title{data_normalization}
\usage{
data_normalization(y, type = "gauss")
}
\arguments{
\item{y}{time series to segment}

\item{type}{the model: "gauss", "exp", "poisson", "geom", "bern", "binom", "negbin"}
}
\value{
the same time series `y` after normalization
}
\description{
Normalizing data for use in algorithm
}
\examples{
data_normalization(rnorm(100), type = "gauss")
}
