% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataGenerator}
\alias{dataGenerator}
\title{dataGenerator}
\usage{
dataGenerator(chpts = 100, parameters = 0, sdNoise = 1, type = "gauss")
}
\arguments{
\item{chpts}{a vector of increasing change-point indices}

\item{parameters}{vector of successive segment parameters}

\item{sdNoise}{standard deviation for the noise parameter (if available)}

\item{type}{the model: gauss or poisson}
}
\value{
a vector of simulated time series
}
\description{
Generating data for change-point detection for univariate time series
}
\examples{
myData <- dataGenerator(chpts = c(30,100,120), parameter = c(0,5,0))
}
