% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataGenerator.R
\name{dataGenerator_MultiD}
\alias{dataGenerator_MultiD}
\title{dataGenerator_MultiD}
\usage{
dataGenerator_MultiD(
  chpts = 100,
  parameters = 0,
  sdNoise = 1,
  nbTrials = 10,
  nbFailures = 10,
  type = "gauss"
)
}
\arguments{
\item{chpts}{a vector of increasing change-point indices}

\item{parameters}{matrix of successive segment parameters}

\item{sdNoise}{standard deviation for the noise parameter (type gauss)}

\item{nbTrials}{number of trials (type binom)}

\item{nbFailures}{number of failures (type negbin)}

\item{type}{the model: gauss, poisson}
}
\value{
a vector of simulated time series
}
\description{
Generating data for multiple change-point detection in multivariate time series binding together 1D time-series from dataGenerator1D
}
\examples{
myData <- dataGenerator_MultiD(chpts = c(50,100), parameter = c(0,1), type = "gauss")
}
