% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust.R
\name{dust.MD}
\alias{dust.MD}
\title{Run the MD Dust Change Point Detection}
\usage{
dust.MD(
  data = data,
  penalty = 2 * nrow(data) * log(length(data)),
  constraints = nrow(data),
  model = "gauss",
  method = "fastest",
  alpha = 1e-09,
  nbLoops = 10
)
}
\arguments{
\item{data}{A numeric matrix. The time series data on which change point detection is performed. Each row should be a different dimension.}

\item{penalty}{A numeric value. The penalty applied for adding a new change point. By default, it is set to \code{2 nrow(data) log(length(data))}.}

\item{constraints}{An integer. The maximum number of constraints to be considered in the DUST pruning test.}

\item{model}{A character string. Specifies the model used for change point detection. Default is "gauss". Possible values could include "gauss", "poisson", "exp", "geom", "binom", "negbin", "variance".}

\item{method}{A character string. Specifies the method to be used in the detection. Default is "fastest". Different segmentation methods may be specified.}

\item{alpha}{A numeric value. For randomness level. Default is \code{1e-9}.}

\item{nbLoops}{An integer. The number of loops to run in the max dual optimization algorithm. Default is 10.}
}
\value{
A list containing the change points detected by the DUST algorithm.
}
\description{
This function performs change point detection on one-dimensional data using the DUST algorithm.
}
\examples{
data <- matrix(rnorm(1000 * 2), nrow = 2, ncol = 1000)
result <- dust.MD(data = data)
}
