% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DustPartitioner.R
\name{dust.partitioner.1D}
\alias{dust.partitioner.1D}
\title{dust.partitioner.1D}
\usage{
dust.partitioner.1D(
  model = "gauss",
  method = "fastest",
  alpha = 1e-09,
  nbLoops = 10
)
}
\arguments{
\item{model}{the underlying model of the data. default is "gauss".
available models: c("gauss", "poisson", "exp", "geom", "bern", "binom", "negbin")}

\item{method}{the method for handling the indices and pruning tests in the algorithm.
default is "fastest", which returns the fastest method available on input model. Others are: randIndex_Eval0 to 5 and detIndex_Eval0 to 5}

\item{alpha}{controls the randomness of the random methods. for computational efficiency purposes, a vector of random values is generated upon initializing the partitioner object.
the lower the value of alpha, the larger this vector, meaning a selection of indices closer to "true" randomness.}

\item{nbLoops}{number of iterations in the optimization algorithm for maximizing the dual function}
}
\value{
a DUST partitioner object that provides methods : fit, for fitting the data; compute, once fit has been called, for computing the optimal partition of the data; get_partition, for retrieving the optimal partition once it has been computed; and quick, a wrapper for the 3 methods.
}
\description{
Generates a DUST partitioner
}
\examples{
dust.partitioner.1D()
}
