% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DustPartitioner.R
\name{dust.partitioner.meanVar}
\alias{dust.partitioner.meanVar}
\title{dust.partitioner.meanVar}
\usage{
dust.partitioner.meanVar(method = "fastest", alpha = 1e-09, nbLoops = 10)
}
\arguments{
\item{method}{the method for handling the indices and pruning tests in the algorithm. defaults to "fastest", which returns the fastest method available on input model. available methods: "randIndex_randEval" prunes indices based on a random dual test, where the value of the dual is evaluated at a random point, and the dual is defined by a random index; "randIndex_detEval", where the dual is defined by a random index but evaluated at its maximum.}

\item{alpha}{controls the randomness of the random methods. for computational efficiency purposes, a vector of random values is generated upon initializing the partitioner object. the lower the value of alpha, the larger this vector, meaning a selection of indices closer to "true" randomness.}

\item{nbLoops}{number of iteration in the optimization algorithm for maximizing the dual function}
}
\value{
a DUST partitioner object that provides methods : fit, for fitting the data; compute, once fit has been called, for computing the optimal partition of the data; get_partition, for retrieving the optimal partition once it has been computed; and quick, a wrapper for the other 3 methods.
}
\description{
Generates a DUST partitioner
}
\examples{
dust.partitioner.meanVar()
}
