% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust_R.R
\name{dust_R}
\alias{dust_R}
\title{dust_R}
\usage{
dust_R(data, penalty, type = "gauss")
}
\arguments{
\item{data}{a vector of data (univariate)}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector `\code{nb} saving the number of non-pruned elements at each iteration
}
\description{
DUST algorithm for univariate time-series (with different possible data models)
}
\examples{
dust_R(dataGenerator_1D(chpts = c(50,200,400), parameters = c(0,1,0), type = "gauss"), 2*log(400))
}
