% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust_R.R
\name{dust_R_1D}
\alias{dust_R_1D}
\title{dust_R_1D}
\usage{
dust_R_1D(data, penalty, type = "gauss", pruningOpt = 1)
}
\arguments{
\item{data}{a vector of data to segment generated by \code{dataGenerator_1D}}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}}

\item{pruningOpt}{the pruning option to use. 1 = dust(uniform), 2 = dust(uniform) + PELT}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector `\code{nb} saving the number of non-pruned elements at each iteration
}
\description{
DUST algorithm for univariate time-series (with different possible data models)
}
\examples{
n <- 100
type <- "gauss"
data <- dataGenerator_1D(chpts = c(n/2,n), parameters = c(0,2), type = type)
dust_R_1D(data, 2*log(n), type = type)
type <- "poisson"
data <- dataGenerator_1D(chpts = c(n/2,n), parameters = c(5,2), type = type)
dust_R_1D(data, 2*log(n), type = type)
type <- "exp"
data <- dataGenerator_1D(chpts = c(n/2,n), parameters = c(5,1), type = type)
dust_R_1D(data, 2*log(n), type = type)
type <- "binom"
data <- dataGenerator_1D(chpts = c(n/2,n), parameters = c(0.2,0.7), type = type)
dust_R_1D(data/10, 2*log(n), type = type, pruningOpt = 2)
}
