% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_R.R
\name{OP_R}
\alias{OP_R}
\title{OP_R}
\usage{
OP_R(data, penalty, type = "gauss")
}
\arguments{
\item{data}{a vector of data}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: gauss, poisson, exp}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector `nb` saving the number of non-pruned elements at each iteration
}
\description{
OP algorithm for univariate time-series (with different possible data models)
}
\examples{
OP_R(dataGenerator_1D(chpts = c(50,200,400), parameters = c(0,1,0), type = "gauss"), 2*log(400))
}
