% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_R.R
\name{OP_R_2Dquad}
\alias{OP_R_2Dquad}
\title{OP_R_2Dquad}
\usage{
OP_R_2Dquad(data, penalty, type = "meanVar")
}
\arguments{
\item{data}{a vector of data (type \code{"meanVar"}) generated by \code{dataGenerator_MV} or a data frame of data of dimension n x 2 (type \code{"regression"}) generated by \code{dataGenerator_Reg}.}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"meanVar"}, \code{"regression"}}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector \code{costQ} saving the optimal cost at each time step
}
\description{
OP algorithm for bivariate time-series (with different possible links)
}
\examples{
 OP_R_2Dquad(dataGenerator_MV(c(50,100), c(0,1), c(0.5,0.6)), 4*log(100))
 OP_R_2Dquad(dataGenerator_Reg(c(40,90), c(2,-1),  c(-1,2), c(1,2)), log(100))
}
