% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OP_R.R
\name{OP_R_MultiD}
\alias{OP_R_MultiD}
\title{OP_R_MultiD}
\usage{
OP_R_MultiD(data, penalty, type = "gauss")
}
\arguments{
\item{data}{a matrix of data to segment generated by \code{dataGenerator_MultiD}}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector \code{costQ} saving the optimal cost at each time step
}
\description{
OP algorithm for univariate time-series (with different possible data models)
}
