% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalization.R
\name{sdDiff}
\alias{sdDiff}
\title{sdDiff}
\usage{
sdDiff(y, method = "HALL")
}
\arguments{
\item{y}{Univariate time series}

\item{method}{Three available methods: \code{"HALL"}, \code{"MAD"} and \code{"SD"}}
}
\value{
a value equal to the estimated standard deviation
}
\description{
sdDiff is a function based on the difference operator (or difference order for HALL method) estimating the time-series standard deviation in Gaussian model
The estimation is efficient for time-series with constant standard deviation and multiple changes in mean.
Three estimators are available:
\itemize{
  \item \code{HALL} : the so-called HALL-estimator of order 3. For more details see: \emph{(1990) Asymptotically optimal difference-based estimation of variance in nonparametric regression. Authors: Hall, Peter and Kay, JW and Titterinton, DM. Biometrika, pages 521--528}
  \item \code{MAD} : the Median Absolute Deviation estimator computed on \code{diff(x)/sqrt(2)} with x the vector of data-points
  \item \code{SD} : the standard deviation estimator (function sd) computed on \code{diff(x)/sqrt(2)} with x the vector of data-points
}
}
\examples{
data <- dataGenerator_1D(chpts = c(100, 200, 300), parameters = c(0,1,0), sdNoise = 2)
sdDiff(data, method = "HALL")
sdDiff(data, method = "MAD")
sdDiff(data, method = "SD")
}
