% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{data_normalization}
\alias{data_normalization}
\title{Data Normalization Function}
\usage{
data_normalization(y, type = "gauss")
}
\arguments{
\item{y}{A numeric vector representing the time series to be normalized and then segmented.}

\item{type}{A string specifying the model type for normalization.
The available options are "gauss", "exp", "poisson", "geom", "bern", "binom", and "negbin".
The default is "gauss".}
}
\value{
A numeric vector that is the normalized version of the input time series `y`.
}
\description{
Normalizes the input time series data `y` according to the specified `type`.
The normalization process depends on the statistical model type, which can be one of the following:
"gauss" (Gaussian/normal distribution), "exp" (exponential distribution),
"poisson" (Poisson distribution), "geom" (geometric distribution),
"bern" (Bernoulli distribution), "binom" (binomial distribution), or
"negbin" (negative binomial distribution).
}
\examples{
# Normalize a random time series using the Gaussian model
normalized_y <- data_normalization(rnorm(100), type = "gauss")

# Normalize using the Poisson model
normalized_y <- data_normalization(rpois(100, lambda = 3), type = "poisson")

# Normalize using the Exponential model
normalized_y <- data_normalization(rexp(100), type = "exp")

}
