% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust.R
\name{dust.1D}
\alias{dust.1D}
\title{Run the 1D Dust Change Point Detection}
\usage{
dust.1D(
  data = data,
  penalty = 2 * log(length(data)),
  model = "gauss",
  method = "fastest",
  alpha = 1e-09,
  nbLoops = 10
)
}
\arguments{
\item{data}{A numeric vector. The time series data on which change point detection is performed.}

\item{penalty}{A numeric value. The penalty applied for adding a new change point. By default, it is set to \code{2 * log(length(data))}.}

\item{model}{A character string. Specifies the model used for change point detection. Default is "gauss". Possible values could include "gauss", "poisson", "exp", "geom", "binom", "negbin".}

\item{method}{A character string. Specifies the method to be used in the detection. Default is "fastest". Different segmentation methods may be specified.}

\item{alpha}{A numeric value. For randomness level. Default is \code{1e-9}.}

\item{nbLoops}{An integer. The number of loops to run in the max dual optimization algorithm. Default is 10.}
}
\value{
A list containing the change points detected by the DUST algorithm.
}
\description{
This function performs change point detection on one-dimensional data using the DUST algorithm.
}
\examples{
data <- rnorm(1000)
result <- dust.1D(data = data)
}
