% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dust_R}
\alias{dust_R}
\title{dust_R}
\usage{
dust_R(data, penalty, type = "gauss")
}
\arguments{
\item{data}{a vector of data}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: gauss, poisson, exp}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector `nb` saving the number of non-pruned elements at each iteration
}
\description{
DUST algorithm for univariate time-series (with different possible data models)
}
\examples{
dust_R(dataGenerator(chpts = c(30,100,120), means = c(0,1,0), type = "gauss"))
}
