% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_OP.R
\name{OP_R_MultiD}
\alias{OP_R_MultiD}
\title{OP_R_MultiD}
\usage{
OP_R_MultiD(data, penalty = 2 * nrow(data) * log(ncol(data)), type = "gauss")
}
\arguments{
\item{data}{a matrix of data to segment generated by \code{dataGenerator_MultiD}}

\item{penalty}{penalty value (non-negative)}

\item{type}{type of cost to use: \code{"gauss"}, \code{"exp"}, \code{"poisson"}, \code{"geom"}, \code{"bern"}, \code{"binom"}, \code{"negbin"}}
}
\value{
a list with the change-point elements (each last index of each segment) and a vector \code{costQ} saving the optimal cost at each time step
}
\description{
OP algorithm for multivariate time-series (with different possible data models)
}
\examples{
data <- dataGenerator_MultiD(chpts = c(100,200),
                             parameters = data.frame(ts1 = c(0,-2), ts2 = c(0,1)))
OP_R_MultiD(data)$changepoints

data <- dataGenerator_MultiD(chpts = c(50,100,150),
   parameters = data.frame(ts1 = c(0.4,0.3,0.5), ts2 = c(0.5,0.6,0.5), ts3 = c(0.3,0.4,0.6)),
   nbTrials = c(30,10,100), type = "binom")
OP_R_MultiD(data, 50*log(150))$changepoints
}
